изменена логика подсчета общего числа документов
git-svn-id: http://svn.datateh.ru/alfrescoecm/Business%20Platform/trunk@4760 697f8a5b-bb4b-4f4c-ab90-ff1ad96ece97

+++ lecm-documents/api/src/main/java/ru/it/lecm/documents/beans/DocumentService.java
@@ -149,7 +149,7 @@ public interface DocumentService {
      * @param sortDefinition - набор полей для сортировки
      * @return Integer - количество документов
      */
-    public Integer getAmountDocumentsByFilter(List<QName> docTypes, List<String> paths, List<String> statuses, String filterQuery, List<SortDefinition> sortDefinition);
+    public Long getAmountDocumentsByFilter(List<QName> docTypes, List<String> paths, List<String> statuses, String filterQuery, List<SortDefinition> sortDefinition);
 
     /**
      * Метод для получения папки с черновиками для заданного типа
